+++ com.unity.render-pipelines.high-definition/Editor/ShaderGraph/HDFields.cs
@@ -175,6 +175,7 @@ namespace UnityEditor.Rendering.HighDefinition
             public static FieldDescriptor Fabric = new FieldDescriptor(kSubShader, "Fabric SubShader", "");
             public static FieldDescriptor StackLit = new FieldDescriptor(kSubShader, "StackLit SubShader", "");
             public static FieldDescriptor Unlit = new FieldDescriptor(kSubShader, "Unlit SubShader", "");
+            public static FieldDescriptor Hair = new FieldDescriptor(kSubShader, "Hair SubShader", "");
         }
 
 #endregion