+++ com.unity.render-pipelines.high-definition/Editor/ShaderGraph/HDSubShaders.cs
@@ -191,6 +191,20 @@ namespace UnityEditor.Rendering.HighDefinition.ShaderGraph
             },
         };
 
+        public static SubShaderDescriptor HairRaytracing = new SubShaderDescriptor()
+        {
+            pipelineTag = HDRenderPipeline.k_ShaderTagName,
+            generatesPreview = false,
+            passes = new PassCollection
+            {
+                { HDPasses.HairRayTracing.Indirect, new FieldCondition(Fields.IsPreview, false) },
+                { HDPasses.HairRayTracing.Visibility, new FieldCondition(Fields.IsPreview, false) },
+                { HDPasses.HairRayTracing.Forward, new FieldCondition(Fields.IsPreview, false) },
+                { HDPasses.HairRayTracing.GBuffer, new FieldCondition(Fields.IsPreview, false) },
+                { HDPasses.HairRayTracing.SubSurface, new FieldCondition(Fields.IsPreview, false) },
+            },
+        };
+
         public static SubShaderDescriptor HDLitRaytracing = new SubShaderDescriptor()
         {
             pipelineTag = HDRenderPipeline.k_ShaderTagName,